chore: move notification urls / views to api
