+++ backend/notifications/urls.py
@@ -1,6 +1,1 @@
-from django.urls import path
-from notifications.views import NotificationView
-
-urlpatterns = [
-    path("tmp/<str:user_id>/", NotificationView.as_view(), name="notification-detail"),
-]
+urlpatterns = []
